full app working

+++ lib/CustomWidgets/calandar_widget.dart
@@ -1,8 +1,10 @@
 // import 'package:crc_app/Api/api.dart';
+import 'package:crc_app/Api/api.dart';
 import 'package:crc_app/CustomWidgets/snack_bar.dart';
 import 'package:crc_app/main.dart';
 import 'package:crc_app/styles.dart';
 import 'package:crc_app/userStatusProvider/user_and_event_provider.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 // import 'package:http/http.dart';
 import 'package:provider/provider.dart';
@@ -75,6 +77,7 @@ class _CalandarWidgetState extends State<CalandarWidget> {
       todayHighlightColor: prussianBlue,
       onTap: (CalendarTapDetails details) {
         //TODO show dialogue box of event info here
+        debugPrint("${details.appointments}");
         String currentId = details.appointments![0].id;
         Map<String, dynamic> currentEvent = findCurrentEvent(currentId);
         // print(currentEvent);
@@ -98,51 +101,55 @@ class _CalandarWidgetState extends State<CalandarWidget> {
 
   void createDisplayEventsMap() {
     debugPrint("entered displayevents function"); //TODO remove this
+    // print(widget.eventsMap);
     for (Map<String, dynamic> event in widget.eventsMap) {
       Map<String, dynamic> displayEvent = {};
-      // DateTime eventDate = DateTime.parse(event["eventDate"]);
-      // int startTime = int.parse(event["startTime"]);
-      // int endTime = int.parse(event["endTime"]);
-      displayEvent["id"] = event["id"];
-      displayEvent["title"] = event["eventName"];
-      // displayEvent["startTime"] =
+      // DateTime eventDate = DateTime.parse(event["Date"]);
+      // int startTime = int.parse(event["BookedFrom"]);
+      // int endTime = int.parse(event["BookedTill"]);
+      displayEvent["_id"] = event["_id"];
+      debugPrint(event["_id"]);
+      displayEvent["title"] = event["EventName"];
+      // displayEvent["BookedFrom"] =
       //     DateTime(eventDate.year, eventDate.month, eventDate.day, startTime);
-      // displayEvent["endTime"] =
+      // displayEvent["BookedTill"] =
       //     DateTime(eventDate.year, eventDate.month, eventDate.day, endTime);
       //TODO added below two lines
-      displayEvent["startTime"] = DateTime.parse(event["startTime"]);
-      displayEvent["endTime"] = DateTime.parse(event["endTime"]);
-      if (event["status"] == "Booked") {
+      displayEvent["BookedFrom"] = DateTime.parse(event["BookedFrom"]);
+      displayEvent["BookedTill"] = DateTime.parse(event["BookedTill"]);
+      if (event["Status"] == "booked") {
         displayEvent["color"] = Colors.orangeAccent;
-      } else if (event["status"] == "inUse") {
+      } else if (event["Status"] == "inUse") {
         displayEvent["color"] = Colors.redAccent;
       } else {
-        displayEvent["color"] = Color(0xff69d1c5);
+        displayEvent["color"] = const Color(0xff69d1c5);
       }
       diaplayEventsList.add(displayEvent);
     }
   }
 
   Map<String, dynamic> findCurrentEvent(String currentId) {
-    Map<String, dynamic> return_event = {};
+    Map<String, dynamic> returnEvent = {};
     bool found = false;
     int length = widget.eventsMap.length;
     for (int i = 0; i < length && !found; i++) {
-      if (widget.eventsMap[i]["id"] == currentId) {
-        return_event = widget.eventsMap[i];
+      if (widget.eventsMap[i]["_id"] == currentId) {
+        returnEvent = widget.eventsMap[i];
       }
     }
-    return return_event;
+    return returnEvent;
   }
 }
 
 class EventDataSource extends CalendarDataSource {
   EventDataSource(List<Map<String, dynamic>> eventsMap) {
     appointments = eventsMap.map((eventMap) {
+      //!error here
+      print("id is: ${eventMap["_id"]}");
       return Appointment(
-        id: eventMap["id"],
-        startTime: eventMap["startTime"],
-        endTime: eventMap["endTime"],
+        id: "${eventMap["_id"]}",
+        startTime: eventMap["BookedFrom"],
+        endTime: eventMap["BookedTill"],
         subject: eventMap["title"],
         color: eventMap["color"], // Customize color as needed
       );
@@ -169,33 +176,39 @@ class _EventDialogBoxState extends State<EventDialogBox> {
   void dispose() {
     // TODO: implement dispose
     super.dispose();
+    print("Entered event dialog box");
     otpTextField.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    print(widget.currentEventMap);
+    if (kDebugMode) {
+      print(widget.currentEventMap);
+    }
     // String user = "Guest";
-    DateTime currentEventDate =
-        DateTime.parse(widget.currentEventMap["eventDate"]);
-    // int currentStartTime = int.parse(widget.currentEventMap["startTime"]);
-    // int currentEndTime = int.parse(widget.currentEventMap["endTime"]);
+    DateTime currentEventDate = DateTime.parse(widget.currentEventMap["Date"]);
+    // print(widget.currentEventMap);
+    // int currentStartTime = int.parse(widget.currentEventMap["BookedFrom"]);
+    // int currentEndTime = int.parse(widget.currentEventMap["BookedTill"]);
     DateTime currentStartTime =
-        DateTime.parse(widget.currentEventMap["startTime"]);
-    DateTime currentEndTime = DateTime.parse(widget.currentEventMap["endTime"]);
+        DateTime.parse(widget.currentEventMap["BookedFrom"]);
+    DateTime currentEndTime =
+        DateTime.parse(widget.currentEventMap["BookedTill"]);
     //TODO if not using currentstarttime and end time as datetime, remove the .hour from the below code
     String currentStartTimeString =
         "${currentStartTime.hour <= 12 ? currentStartTime.hour : currentStartTime.hour - 12}:00 ${currentStartTime.hour < 12 ? "AM" : "PM"}";
     String currentEndTimeString =
         "${currentEndTime.hour <= 12 ? currentEndTime.hour : currentEndTime.hour - 12}:00 ${currentEndTime.hour < 12 ? "AM" : "PM"}";
     String statusString;
-    if (widget.currentEventMap["status"] == "Booked") {
-      statusString = "Booked";
-    } else if (widget.currentEventMap["status"] == "inUse") {
+    if (widget.currentEventMap["Status"] == "booked") {
+      statusString = "booked";
+    } else if (widget.currentEventMap["Status"] == "inUse") {
       statusString = "In Use";
     } else {
       statusString = "Free";
     }
+    String currentId = widget.currentEventMap["_id"];
+    String currentOTP = currentId.substring(currentId.length - 4);
     return AlertDialog(
       backgroundColor: backgroundColor,
       titlePadding: const EdgeInsets.only(left: 0, right: 0, top: 10),
@@ -211,8 +224,7 @@ class _EventDialogBoxState extends State<EventDialogBox> {
                   visible: widget.isAdmin,
                   child: IconButton(
                       onPressed: () async {
-                        bool eventDeleted =
-                            await deleteEvent(widget.currentEventMap["id"]);
+                        bool eventDeleted = await deleteEvent(currentId);
                         if (eventDeleted) {
                           Navigator.pop(context);
                         }
@@ -228,7 +240,7 @@ class _EventDialogBoxState extends State<EventDialogBox> {
               flex: 2,
               child: Center(
                 child: Text(
-                  "${widget.currentEventMap["eventName"]}",
+                  "${widget.currentEventMap["EventName"]}",
                   style: TextStyle(
                       fontSize: 20,
                       fontWeight: FontWeight.bold,
@@ -259,11 +271,13 @@ class _EventDialogBoxState extends State<EventDialogBox> {
         children: [
           TextRichEventDialog(
             heading: "Organiser",
-            value: widget.currentEventMap["organiserName"],
+            value: widget.currentEventMap["OrganiserName"],
           ),
           TextRichEventDialog(
             heading: "Mobile No.",
-            value: widget.currentEventMap["mobileNumber"],
+            //todo change to the actual mobile number from map
+            //widget.currentEventMap["MobileNumber"]
+            value: "947808844",
           ),
           TextRichEventDialog(
               heading: "Event Date",
@@ -273,11 +287,14 @@ class _EventDialogBoxState extends State<EventDialogBox> {
               heading: "Start Time", value: currentStartTimeString),
           TextRichEventDialog(heading: "End Time", value: currentEndTimeString),
           TextRichEventDialog(heading: "Status", value: statusString),
-          if (!widget.isAdmin && widget.currentEventMap["status"] != "free")
+          if (widget.isAdmin)
+            TextRichEventDialog(heading: "OTP", value: currentOTP),
+          if (!widget.isAdmin && widget.currentEventMap["Status"] != "free")
             const SizedBox(
               height: 10,
             ),
-          if (!widget.isAdmin && widget.currentEventMap["status"] == "Booked")
+          //! change to !widget.isAdmin below
+          if (!widget.isAdmin && widget.currentEventMap["Status"] == "booked")
             Form(
               key: _formKey,
               child: Column(
@@ -313,8 +330,9 @@ class _EventDialogBoxState extends State<EventDialogBox> {
                       ),
                       onPressed: () async {
                         bool otpVerified = await verifyOtp(
-                            widget.currentEventMap["id"],
-                            otpTextField.text.trim());
+                            widget.currentEventMap["_id"],
+                            otpTextField.text.trim(),
+                            currentOTP);
                         if (otpVerified) {
                           Navigator.pop(context);
                         }
@@ -331,14 +349,15 @@ class _EventDialogBoxState extends State<EventDialogBox> {
                 ],
               ),
             ),
-          if (!widget.isAdmin && widget.currentEventMap["status"] == "inUse")
+
+          if (!widget.isAdmin && widget.currentEventMap["Status"] == "inUse")
             SizedBox(
               width: double.infinity,
               height: MediaQuery.of(context).size.height * 0.06,
               child: ElevatedButton(
                 onPressed: () async {
                   bool keysReturned =
-                      await returnKeys(widget.currentEventMap["id"]);
+                      await returnKeys(widget.currentEventMap["_id"]);
                   if (keysReturned) {
                     Navigator.pop(context);
                   }
@@ -367,38 +386,52 @@ class _EventDialogBoxState extends State<EventDialogBox> {
   }
 
   Future<bool> deleteEvent(String id) async {
-    bool deleted = false;
+    bool eventDeleted;
     String snakbarText = "Failed to delete";
-    print(id);
+    if (kDebugMode) {
+      print(id);
+    }
     try {
       //TODO delete here
-      // ApiService().deleteData(id);
-      final provider =
-          navigatorKey.currentState!.context.read<UserStatusProvider>();
-      provider.deleteEventDataById(id);
-      deleted = true;
-      snakbarText = "Event deleted";
+      eventDeleted = await ApiService().deleteData(id);
+      if (eventDeleted) {
+        final provider =
+            navigatorKey.currentState!.context.read<UserStatusProvider>();
+        provider.deleteEventDataById(id);
+        debugPrint("Deleted successfully");
+        snakbarText = "Event deleted";
+      }
     } catch (e) {
       snakbarText = "NetWork Error";
-      print(e);
+      eventDeleted = false;
+      if (kDebugMode) {
+        print(e);
+      }
     }
-    ScaffoldMessenger.of(context).showSnackBar(
-      customSnackBar(snakbarText),
-    );
-    return deleted;
+    if (eventDeleted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        customSnackBar(snakbarText),
+      );
+    }
+    return eventDeleted;
   }
 
-  Future<bool> verifyOtp(String id, String EnteredOtp) async {
+  Future<bool> verifyOtp(String id, String enteredOtp, String actualOtp) async {
     bool verified = false;
     String snakbarText = "Wrong Otp";
     //TODO check enteered otp
-    if (EnteredOtp == "1821") {
+    if (enteredOtp == actualOtp) {
       try {
-        final provider =
-            navigatorKey.currentState!.context.read<UserStatusProvider>();
-        provider.updateEventStatusById(id, "inUse");
-        verified = true;
-        snakbarText = "Otp Verified";
+        verified = await ApiService().updateBookingStatus(id, "inUse");
+        if (verified) {
+          final provider =
+              navigatorKey.currentState!.context.read<UserStatusProvider>();
+          provider.updateEventStatusById(id, "inUse");
+          verified = true;
+          snakbarText = "Otp Verified";
+        } else {
+          snakbarText = "NetWork Error";
+        }
       } catch (e) {
         snakbarText = "NetWork Error";
         debugPrint("$e");
@@ -414,11 +447,14 @@ class _EventDialogBoxState extends State<EventDialogBox> {
     bool keysReturned = false;
     String snakbarText = "Keys Not Returned";
     try {
-      final provider =
-          navigatorKey.currentState!.context.read<UserStatusProvider>();
-      provider.updateEventStatusById(id, "free");
-      snakbarText = "Keys Returned";
-      keysReturned = true;
+      keysReturned = await ApiService().updateBookingStatus(id, "free");
+      if (keysReturned) {
+        final provider =
+            navigatorKey.currentState!.context.read<UserStatusProvider>();
+        provider.updateEventStatusById(id, "free");
+        snakbarText = "Keys Returned";
+        keysReturned = true;
+      }
     } catch (e) {
       snakbarText = "NetWork Error";
       debugPrint("$e");
